Refactor Padding tests
